# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CanFactory API",
    "version": "1.0.0",
    "description": "Local parametric STL generation. Dimensions are millimetres. Generated jobs and files expire one hour after creation; there are no saved user designs."
  },
...
  "paths": {
...
    "/api/v1/models/{id}": {
      "get": {
        "operationId": "getModel",
        "summary": "Get a model and its editor schema",
        "tags": [
          "Models"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[a-zA-Z0-9-]+$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "version",
                    "title",
                    "description",
                    "attribution",
                    "license",
                    "licenseUrl",
                    "controls",
                    "defaults",
                    "parameterSchema",
                    "referenceUrl"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "attribution": {
                      "type": "string"
                    },
                    "license": {
                      "type": "string"
                    },
                    "licenseUrl": {
                      "type": "string"
                    },
                    "controls": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "key",
                          "label",
                          "description",
                          "kind",
                          "group",
                          "unit",
                          "default",
                          "minimum",
                          "maximum",
                          "step",
                          "enabledWhen"
                        ],
                        "properties": {
                          "key": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "kind": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "number"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "boolean"
                                ]
                              }
                            ]
                          },
                          "group": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "basic"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "advanced"
                                ]
                              }
                            ]
                          },
                          "unit": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "mm"
                                ]
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "default": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "boolean"
                              }
                            ]
                          },
                          "minimum": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "maximum": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "step": {
                            "anyOf": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "enabledWhen": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        },
                        "additionalProperties": false
                      }
                    },
                    "defaults": {
                      "type": "object",
                      "additionalProperties": {
                        "anyOf": [
                          {
                            "type": "number"
                          },
                          {
                            "type": "boolean"
                          }
                        ]
                      }
                    },
                    "parameterSchema": {
                      "type": "object",
                      "description": "JSON Schema for this model’s parameter object."
                    },
                    "referenceUrl": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "issues"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Stable machine-readable error code.",
                      "example": "INVALID_PARAMETERS"
                    },
                    "message": {
                      "type": "string",
                      "description": "Actionable human-readable explanation."
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "field",
                          "message"
                        ],
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": false
                      }
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/{id}/reference.stl": {
      "get": {
        "operationId": "getReferenceStl",
        "summary": "Inspect the original supplied STL",
        "tags": [
          "Models"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[a-zA-Z0-9-]+$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Original STL bytes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "description": "Original STL bytes."
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "issues"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Stable machine-readable error code.",
                      "example": "INVALID_PARAMETERS"
                    },
                    "message": {
                      "type": "string",
                      "description": "Actionable human-readable explanation."
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "field",
                          "message"
                        ],
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": false
                      }
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/renders/{id}": {
      "get": {
        "operationId": "getRender",
        "summary": "Poll a render job",
        "tags": [
          "Renders"
        ],
        "description": "410 means this temporary render is expired or no longer available. Resubmit current settings to regenerate it.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[a-zA-Z0-9-]+$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "modelId",
                    "modelVersion",
                    "status",
                    "createdAt",
                    "expiresAt",
                    "slotCount",
                    "artifact",
                    "error"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "modelId": {
                      "type": "string"
                    },
                    "modelVersion": {
                      "type": "string"
                    },
                    "status": {
                      "enum": [
                        "queued",
                        "running",
                        "succeeded",
                        "failed"
                      ]
                    },
                    "createdAt": {
                      "type": "integer"
                    },
                    "expiresAt": {
                      "type": "integer"
                    },
                    "slotCount": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "artifact": {
                      "anyOf": [
                        {
                          "type": "object",
                          "required": [
                            "url",
                            "sha256",
                            "bytes",
                            "triangles",
                            "dimensions",
                            "volume"
                          ],
                          "properties": {
                            "url": {
                              "type": "string",
                              "description": "The same bytes are used for preview and download. Add ?download=true for attachment disposition."
                            },
                            "sha256": {
                              "type": "string"
                            },
                            "bytes": {
                              "type": "integer"
                            },
                            "triangles": {
                              "type": "integer"
                            },
                            "dimensions": {
                              "type": "object",
                              "required": [
                                "x",
                                "y",
                                "z"
                              ],
                              "properties": {
                                "x": {
                                  "type": "number"
                                },
                                "y": {
                                  "type": "number"
                                },
                                "z": {
                                  "type": "number"
                                }
                              },
                              "additionalProperties": false,
                              "description": "Axis-aligned dimensions in millimetres, including brim and handles."
                            },
                            "volume": {
                              "type": "number",
                              "description": "Enclosed material volume in cubic millimetres."
                            }
                          },
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "error": {
                      "anyOf": [
                        {
                          "type": "object",
                          "required": [
                            "code",
                            "message",
                            "issues"
                          ],
                          "properties": {
                            "code": {
                              "type": "string",
                              "description": "Stable machine-readable error code.",
                              "examples": [
                                "INVALID_PARAMETERS"
                              ]
                            },
                            "message": {
                              "type": "string",
                              "description": "Actionable human-readable explanation."
                            },
                            "issues": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "required": [
                                  "field",
                                  "message"
                                ],
                                "properties": {
                                  "field": {
                                    "type": "string"
                                  },
                                  "message": {
                                    "type": "string"
                                  }
                                },
                                "additionalProperties": false
                              }
                            }
                          },
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "410": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "issues"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Stable machine-readable error code.",
                      "example": "INVALID_PARAMETERS"
                    },
                    "message": {
                      "type": "string",
                      "description": "Actionable human-readable explanation."
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "field",
                          "message"
                        ],
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": false
                      }
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/renders/{id}/stl": {
      "get": {
        "operationId": "getRenderStl",
        "summary": "View or download the generated STL",
        "tags": [
          "Renders"
        ],
        "description": "Preview and download return identical bytes. Use download=true for attachment disposition. Attribution is recorded in the catalogue and STL header.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "true"
              ]
            },
            "in": "query",
            "name": "download",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[a-zA-Z0-9-]+$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Validated binary STL, measured in mm.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "description": "Validated binary STL, measured in mm."
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "issues"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Stable machine-readable error code.",
                      "example": "INVALID_PARAMETERS"
                    },
                    "message": {
                      "type": "string",
                      "description": "Actionable human-readable explanation."
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "field",
                          "message"
                        ],
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": false
                      }
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "410": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "issues"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "Stable machine-readable error code.",
                      "example": "INVALID_PARAMETERS"
                    },
                    "message": {
                      "type": "string",
                      "description": "Actionable human-readable explanation."
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "field",
                          "message"
                        ],
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": false
                      }
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}